1. Модульные тесты. [#159435]

+++ src/ru/job4j/converter/Converter.java
@@ -1,18 +1,18 @@
 package ru.job4j.converter;
 
 public class Converter {
-    public static float rubToEuro(float value) {
+    public static int rubToEuro(int value) {
         return value / 70;
     }
 
-    public static float rubToUsd(float value) {
+    public static int rubToUsd(int value) {
         return value / 60;
     }
 
     public static void main(String[] args) {
-        float in = 140;
-        float expected = 2;
-        float out = Converter.rubToEuro(in);
+        int in = 140;
+        int expected = 2;
+        int out = Converter.rubToEuro(in);
         boolean passed = expected == out;
         System.out.println("140 rub are 2. Test result: " + passed);
         out = Converter.rubToUsd(in);
